Don't constantly retry an invalid endpoint
If an endpoint isn't enabled on a host (e.g. servicecatalog) previously
this would retry every 5 seconds to connect which is way too frequently.

This adds a time based retry-limit to prevent the collector from
slamming an endpoint that doesn't exist but still allows for a user to
enable an endpoint while the collector is running and the collector to
pick it up without having to restart the collector.